Update git-portfolio schematics
